make tests support both 1.x and 3.x. Lot of work needed to add all the 3.x tests, but that can wait until all is switched

+++ tests/crypto/libcrypto/t_libcrypto.sh
@@ -1,4 +1,4 @@
-# $NetBSD: t_libcrypto.sh,v 1.8 2022/06/10 05:59:12 martin Exp $
+# $NetBSD: t_libcrypto.sh,v 1.9 2023/05/08 17:46:06 christos Exp $
 #
 # Copyright (c) 2008, 2009, 2010 The NetBSD Foundation, Inc.
 # All rights reserved.
@@ -32,7 +32,8 @@ engine_head()
 }
 engine_body()
 {
-	atf_check -o ignore -e ignore "$(atf_get_srcdir)/h_enginetest"
+	atf_check -o ignore -e ignore "$(atf_get_srcdir)/h_enginetest" \
+	    "$(atf_get_srcdir)/d_server.pem"
 }
 
 atf_test_case bn
@@ -68,11 +69,17 @@ threads_head()
 }
 threads_body()
 {
-	$(atf_get_srcdir)/h_threadstest \
-	    -cert $(atf_get_srcdir)/d_server.pem \
-	    -ccert $(atf_get_srcdir)/d_client.pem \
-	    2>&1 | tee out
-	atf_check -s eq:1 -o empty -e empty grep :error: out
+	local s=$(atf_get_srcdir)
+	if [ -f "$s/rsakey.pem" ]; then
+	    atf_check -o ignore -e ignore "$s/h_threadstest" \
+		-config "$s/default.cnf" "$s"
+	else
+		"$s/h_threadstest" \
+		    -cert "$s/d_server.pem" \
+		    -ccert "$s/d_client.pem" \
+		2>&1 | tee out
+		atf_check -s eq:1 -o empty -e empty grep :error: out
+	fi
 }
 
 atf_init_test_cases()
